+++ CHANGELOG.md
@@ -2,7 +2,7 @@
   * `\Sokil\Mongo\Structure\Arrayable` moves to `\Sokil\Mongo\ArrayableInterface`
   * `Structure::$_modifiedFields` and `Structure::$_originalData` set private
   * Document::_data is not deprecated, and replaced with propected property Document::schema.
-  * Documents not alloved to be cloned
+  * Documents not allowed to be cloned
 
 ## 1.15.1 (2016-03-10)
   * Allow delete indexes